requirements-dev.txt
pytest==4.0.1
pytest-django==3.4.4
pytest-pythonpath==0.7.3
python-dateutil==2.7.5
pytest-cov
pylama
codecov

sphinx>=2
sphinx_rtd_theme
sphinxcontrib-images

# TypeError: attrib() got an unexpected keyword argument 'convert'
attrs==19.1.0
